requirements.txt
asgiref==3.5.2
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
distlib==0.3.4
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
filelock==3.7.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
mysqlclient==2.1.0
packaging==21.3
pipenv==2022.5.2
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
